+++ tools/build-package.sh
@@ -75,6 +75,8 @@ find $source_roots -type f \
   ! -path '*/.build-m07/*' \
   ! -path '*/.build-m08/*' \
   ! -path '*/.build-m09/*' \
+  ! -path '*/.build-m10/*' \
+  ! -path '*/.build-m11/*' \
   ! -path '*/.libs/*' \
   ! -path '*/.deps/*' \
   -print | LC_ALL=C sort > "$manifest"